fix datatype from int to long long int

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,18 +10,18 @@
 
 int main(void)
 {
-	int nums[50];
+	long long int nums[50];
 	int i;
 
 	nums[0] = 1;
 	nums[1] = 2;
 
-		printf("%d, %d ", nums[0], nums[1]);
+		printf("%lld, %lld ", nums[0], nums[1]);
 
 	for (i = 2; i < 50; i++)
 	{
 		nums[i] = nums[i - 1] + nums[i - 2];
-		printf("%d", nums[i]);
+		printf("%lld", nums[i]);
 
 		if (i != 49)
 		{
